Refactor repository setup in install script to check for existing configuration before adding the Libre Workspace repository

+++ src/usr/lib/libre-workspace/portal/install.sh
@@ -79,6 +79,11 @@ fi
 
 
 # Setup apt repository if not already done
-curl -1sLf 'https://repo.libre-workspace.org/gpg.key' | sudo gpg --dearmor -o /usr/share/keyrings/libre-workspace-archive-keyring.gpg
-echo "deb [arch=amd64 signed-by=/usr/share/keyrings/libre-workspace-archive-keyring.gpg] https://repo.libre-workspace.org stable main" | sudo tee /etc/apt/sources.list.d/libre-workspace-stable.list > /dev/null
-sudo apt update
+if [ -f /etc/apt/sources.list.d/libre-workspace-stable.list ]; then
+    echo "Libre Workspace repository already set up."
+else
+    echo "Setting up Libre Workspace repository..."
+    curl -1sLf 'https://repo.libre-workspace.org/gpg.key' | sudo gpg --dearmor -o /usr/share/keyrings/libre-workspace-archive-keyring.gpg
+    echo "deb [arch=amd64 signed-by=/usr/share/keyrings/libre-workspace-archive-keyring.gpg] https://repo.libre-workspace.org stable main" | sudo tee /etc/apt/sources.list.d/libre-workspace-stable.list > /dev/null
+    sudo apt update
+fi
